Copyright:
These documents, and the content inside, is hereby given the permission of the owner to be used by anyone obtaining a copy of the document[s].
However, that permission is not granted to change, edit, or delete any part or section from these documents.
If the user wishes to use the document[s] in a paper, essay, etc., the user must give credit to the owner of the document[s].
Any music used we do not have the rights to and anyone pertaining to this game is not allowed to use the music unless owned by that particular person[s].
The name of the game is not given permission by the owner to be used by anyone not pertaining to the creation/release/permission of the game to be used in such a way as to plagiarize the game in any form or way.
Any names of characters, locations, or any other part of the game are purely coincidental and were not used intentionally for plagiarism-type use.
With that being said, any names of characters, locations, or such other original content is not given the right to be used by anyone else and if used by such a person, will be plagiarized.
Any contact needed can be sent to @Aiden.hesser@gmail.com for more information needed or necessary.

Copyright© 2023 by by via @Aiden Hesser